% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annotation_file.R
\name{create_annotation_file}
\alias{create_annotation_file}
\title{Create a template annotation file.}
\usage{
create_annotation_file(title = NULL, annotation_file = NULL)
}
\arguments{
\item{title}{The title within the documentation file.}

\item{annotation_file}{The path to the annotation file to be written.}
}
\description{
Create a template annotation file.
}
\examples{
\dontrun{create_annotation_file()}
}
