% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_overview_toc_chunk.R
\name{create_overview_toc_chunk}
\alias{create_overview_toc_chunk}
\title{Create a table of contents chunk for the overview section.}
\usage{
create_overview_toc_chunk(page_ref)
}
\arguments{
\item{page_ref}{The page_ref ID.}
}
\description{
Create a table of contents chunk for the overview section.
}
\examples{
create_toc_chunk(filename, page_ref)
}
