% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toc_chunk.R
\name{create_toc_chunk}
\alias{create_toc_chunk}
\title{Create a table of contents chunk.}
\usage{
create_toc_chunk(filename, page_ref)
}
\arguments{
\item{filename}{The name of the file for which the chunk is being generated.}

\item{page_ref}{The page_ref ID.}
}
\description{
Create a table of contents chunk.
}
\examples{
create_toc_chunk(filename, page_ref)
}
