% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_it.R
\name{document_it}
\alias{document_it}
\title{Generate documentation for the files of a directory.}
\usage{
document_it(input_directory, output_file, annotation_file = NULL,
  title = NULL)
}
\arguments{
\item{input_directory}{The directory of files to be documented.}

\item{output_file}{The path to the output file that will be generated.}

\item{annotation_file}{The path to the annotation file if present.}

\item{title}{The title of the output document.}
}
\description{
Generate documentation for the files of a directory.
}
\examples{
\dontrun{
document_it(
input_directory = "man",
output_file = "documentation",
annotation_file = NULL
)
}
}
