% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{dr_folders}
\alias{dr_folders}
\title{Default folder names for an export}
\usage{
dr_folders()
}
\value{
A named character vector.
}
\description{
Default folder names for an export
}
\examples{
dr_folders()
}
