% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{dr_output_types}
\alias{dr_output_types}
\title{The output types a collection holds}
\usage{
dr_output_types()
}
\value{
A character vector of output type names.
}
\description{
The output types a collection holds
}
\examples{
dr_output_types()
}
