% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_paragraphs.R
\name{insert_paragraphs}
\alias{insert_paragraphs}
\title{Insert each element of a vector into a document as a string of paragraphs.}
\usage{
insert_paragraphs(denv, vec)
}
\arguments{
\item{denv}{The environment of the input docx object.}

\item{vec}{The character vector to be inserted. Each element should
correspond to a separate paragraph.}
}
\description{
Insert each element of a vector into a document as a string of paragraphs.
}
\examples{
\dontrun{
insert_paragraphs(obj, vec)
}
}
