% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.documenteR_export}
\alias{print.documenteR_export}
\title{Print the result of an export}
\usage{
\method{print}{documenteR_export}(x, ...)
}
\arguments{
\item{x}{A \code{documenteR_export} object, as returned by \code{\link[=save_outputs]{save_outputs()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the result of an export
}
\seealso{
\code{\link[=save_outputs]{save_outputs()}}
}
