% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{documenter}
\alias{documenter}
\alias{documenter_addin}
\title{Document an R object}
\usage{
documenter(objname)

documenter_addin()
}
\arguments{
\item{objname}{A character string naming an R function or data.frame.}
}
\description{
This function takes the name of an object (either an R function or an R
data.frame) and replaces it with skeleton roxygen2 documentation. It is used in the \code{documenter_addin()} function which is the installed R addin.

For \strong{functions}, an empty \code{@param} is generated for each of the funciton's arguments.
For \strong{dataframes}, a full \code{\\description} block is generated from column names
}
\note{
The addin will automatically source the file that the function or data is in.
}
\examples{
documenter("lm")
#' @title FUNCTION_TITLE
#'
#' @description FUNCTION_DESCRIPTION
#'
#' @param formula DESCRIPTION.
#' @param data DESCRIPTION.
#' @param subset DESCRIPTION.
#' @param weights DESCRIPTION.
#' @param na.action DESCRIPTION.
#' @param method DESCRIPTION.
#' @param model DESCRIPTION.
#' @param x DESCRIPTION.
#' @param y DESCRIPTION.
#' @param qr DESCRIPTION.
#' @param singular.ok DESCRIPTION.
#' @param contrasts DESCRIPTION.
#' @param offset DESCRIPTION.
#' @param ... DESCRIPTION.
#'
#' @return RETURN DESCRIPTION
#' @export

documenter("iris")
#' DATASET_TITLE
#'
#' DATASET_DESCRIPTION
#'
#' @format A data frame with 150 rows and 5 variables:
#' \\describe{
#'   \\item{\\code{Sepal.Length}}{double. DESCRIPTION.}
#'   \\item{\\code{Sepal.Width}}{double. DESCRIPTION.}
#'   \\item{\\code{Petal.Length}}{double. DESCRIPTION.}
#'   \\item{\\code{Petal.Width}}{double. DESCRIPTION.}
#'   \\item{\\code{Species}}{integer. DESCRIPTION.}
#' }

}
