% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_annotations.R
\name{extract_annotations}
\alias{extract_annotations}
\title{Extract annotations from a yaml annotation file.}
\usage{
extract_annotations(annotation_yml)
}
\arguments{
\item{annotation_yml}{The yaml object read by document_it.}
}
\description{
Extract annotations from a yaml annotation file.
}
\examples{
extract_annotations(annotation_yml)
}
